// SageNext/sdk-fetch/models/ListItemEntity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Sage Recruitment
 * Simple List API with Nest.js
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListItemEntity
 */
export interface ListItemEntity {
    /**
     * 
     * @type {string}
     * @memberof ListItemEntity
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ListItemEntity
     */
    title: string;
    /**
     * 
     * @type {number}
     * @memberof ListItemEntity
     */
    order: number;
    /**
     * 
     * @type {string}
     * @memberof ListItemEntity
     */
    listId: string;
}

export function ListItemEntityFromJSON(json: any): ListItemEntity {
    return ListItemEntityFromJSONTyped(json, false);
}

export function ListItemEntityFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListItemEntity {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'title': json['title'],
        'order': json['order'],
        'listId': json['listId'],
    };
}

export function ListItemEntityToJSON(value?: ListItemEntity | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'title': value.title,
        'order': value.order,
        'listId': value.listId,
    };
}